refactor(multitenancy): split migration + backfill orphan inferences
Addresses PR review: the inference-derived backfills use INNER JOIN to
tasks, which silently drops rows where inferences.task_id IS NULL.
Deprecated /api/v2/validate_prompt persists exactly those rows; SET NOT
NULL then fails on any DB that's seen that flow.

Fix: rather than papering over the orphans with a default-org fallback,
backfill orphan inferences to the existing __unmapped__ system task
(constants.UNMAPPED_TASK_ID, created in 843e2d3f46d5). Every inference
now has a real task_id; the org chain is unbroken; no special-case logic
in repository code.

Also splits the monolith revision (6a9c711b3de8) into 5 individual
revisions so each step is independently reviewable and revertible:

  mt_create_organizations    organizations table + seed default/system
  mt_tasks_org_id            tasks.org_id NOT NULL (backfill by is_system_task)
  mt_api_keys_org_id         api_keys.org_id NULL (partial index)
  mt_orphan_inferences       UPDATE inferences SET task_id=UNMAPPED WHERE NULL
  mt_denormalize_org_id      org_id on 10 task-scoped sub-resource tables

agentic_annotations also gets a system-org fallback for the rare
"no parent" annotation (NULL trace_id AND NULL continuous_eval_id is
allowed by CK constraints for human annotations).

Verified: full upgrade head, alembic check clean, full downgrade
round-trip, re-upgrade clean.
